misc: added authorization for some resolver

+++ src/graphql/resources/comment/comment.resolvers.ts
@@ -5,6 +5,9 @@ import { Transaction } from "sequelize";
 import { PostInstance } from "../../../models/PostModel";
 import { CommentInstance } from "../../../models/CommentModel";
 import { handleError } from "../../../utils/utils";
+import { compose } from "../../composable/composable.resolver";
+import { authResolvers } from "../../composable/auth.resolver";
+import { AuthUser } from "../../../interfaces/AuthUserInterface";
 
 export const commentResolvers = {
 	Comment: {
@@ -26,18 +29,23 @@ export const commentResolvers = {
 
 	},
 	Mutation: {
-		createComment: (parent, {input}, { db }: { db: DbConnection }, info: GraphQLResolveInfo) => { 
+		createComment: compose(...authResolvers)((parent, {input},  { db, authUser }: { db: DbConnection, authUser: AuthUser }, info: GraphQLResolveInfo) => { 
+			input.user = authUser.id;
 			return db.sequelize.transaction((t: Transaction) => {
 				return db.Comment.create(input, { transaction: t });
 			}).catch(handleError);
-		},
-    updateComment: (parent, {id, input}, { db }: { db: DbConnection }, info: GraphQLResolveInfo) => { 
+		}),
+    updateComment: (parent, {id, input},  { db, authUser }: { db: DbConnection, authUser: AuthUser }, info: GraphQLResolveInfo) => { 
 			id = parseInt(id);
 			return db.sequelize.transaction((t: Transaction) => {
 				return db.Comment.findById(id).then((comment: CommentInstance) => {
 					if (!comment) {
 						throw new Error(`comment ${id} not found`)
 					}
+					if (comment.get('user') !== authUser.id) {
+						throw new Error(`can only update comments that you created`);
+					}
+					input.user = authUser.id;
 					return comment.update(input, {
 						transaction: t
 					}).then((comment: CommentInstance) => {
@@ -46,12 +54,15 @@ export const commentResolvers = {
 				})
 			}).catch(handleError);
 		},
-		deleteComment: (parent, {id}, { db }: { db: DbConnection }, info: GraphQLResolveInfo) => { 
+		deleteComment: (parent, {id},  { db, authUser }: { db: DbConnection, authUser: AuthUser }, info: GraphQLResolveInfo) => { 
 			id = parseInt(id);
 			return db.sequelize.transaction((t: Transaction) => {
 				return db.Comment.findById(id).then((comment: CommentInstance) => {
 					if (!comment) {
 						throw new Error(`comment ${id} not found`)
+					}
+					if (comment.get('user') !== authUser.id) {
+						throw new Error(`can only update comments that you created`);
 					}
 					return comment.destroy({
 						transaction: t
